extracting components for all pages
