+++ app.js
@@ -14,9 +14,6 @@ app.use(require("webpack-hot-middleware")(compiler));
 app.use(express.static('public'));
 app.use('/static', express.static(__dirname + '/static'));
 app.use('/public', express.static(__dirname + '/public'));
-//app.use('/img', express.static(__dirname + '/public/img'));
-//app.use('/css', express.static(__dirname + '/public/css'));
-//app.use('/fonts', express.static(__dirname + '/public/fonts'));
 
 app.get('/', function (req, res) {
   res.sendFile(path.join(__dirname + '/views/index.html'));
