+++ client/index/index.js
@@ -1,0 +1,7 @@
+//rendering of landing page
+
+import React, { PropTypes, Component } from 'react';
+import ReactDOM from 'react-dom';
+import Landing from "./components/Landing.js";
+
+ReactDOM.render(<Landing/>, document.getElementById('app') );
